XGBoost/requirements.txt
spotipy
pandas
numpy
xgboost
scikit-learn
joblib
python-dotenv
